feat: - splitting between contentscript and injectedscript - refactoring the whole injectedScript - seperating between the injectedProvider and the event emitter - enhancing the dapp connection
