+++ package.json
@@ -149,6 +149,7 @@
     "fast-deep-equal": "^3.1.3",
     "link-preview-js": "^3.0.4",
     "loglevel": "^1.8.1",
+    "nanoevents": "^7.0.1",
     "nanoid": "^4.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
