+++ src/injectedScript/providers/EventEmitter.ts
@@ -1,0 +1,23 @@
+import { Emitter, Unsubscribe, createNanoEvents } from "nanoevents";
+
+import { EventHandler, EventName } from "./types";
+
+export class EventEmitter {
+  private emitter: Emitter;
+
+  constructor() {
+    this.emitter = createNanoEvents();
+  }
+
+  on(eventName: EventName, cb: EventHandler): void {
+    this.emitter.on(eventName, cb);
+  }
+
+  emit = (eventName: EventName, payload?: unknown): void => {
+    this.emitter.emit(eventName, payload);
+  };
+
+  cleanListeners(): void {
+    this.emitter.events = {};
+  }
+}
