+++ src/injectedScript/providers/types.ts
@@ -1,0 +1,10 @@
+export interface Approvals {
+  isApproved: boolean;
+  canSkipApprove: boolean;
+}
+
+export type EventHandler = (data: unknown) => void;
+export type EventName = "login" | "identityChanged" | "logout";
+export type Events = {
+  [eventName in EventName]: EventHandler[];
+};
