+++ src/types/contentScript/index.ts
@@ -4,3 +4,10 @@ export interface InjectedMessageData<M = unknown, P = unknown> {
   message: M;
   payload: [string, P];
 }
+
+export interface InjectedProviderRequest {
+  method: string;
+  payload?: unknown;
+  error?: boolean;
+  meta?: Record<string, unknown>;
+}
